Create Linear Search
Problem
You are given an array of integers , you need to find the maximum sum that can be obtained by picking some non-empty subset of the array. If there are many such non-empty subsets, choose the one with the maximum number of elements. Print the maximum sum and the number of elements in the chosen subset.
Sample Input
5
1 2 -4 -2 3
Sample Output
6 3